Updated chat/magic-menu/incpector and header structure.
