+++ src/components/editor-mode/magic-menu/index.tsx
@@ -1,0 +1,3 @@
+
+import { MagicMenu } from './MagicMenu';
+export { MagicMenu };
